all edits

+++ app/src/main/java/ir/guardianapp/guardian_v2/models/Trip.java
@@ -12,7 +12,7 @@ public class Trip {
     private String destinationName;
     private Date startDate;
     private Date endDate;
-    private double distanceInKM;
+    private int distanceInKM;
     private double average;
 
     public Trip(String sourceName, String destinationName, Date startDate, Date endDate, double distanceInKM, double average) {
@@ -20,7 +20,7 @@ public class Trip {
         this.destinationName = destinationName;
         this.startDate = startDate;
         this.endDate = endDate;
-        this.distanceInKM = distanceInKM;
+        this.distanceInKM = (int)distanceInKM;
         this.average = average;
     }
 
@@ -40,7 +40,7 @@ public class Trip {
         return endDate;
     }
 
-    public double getDistanceInKM() {
+    public int getDistanceInKM() {
         return distanceInKM;
     }
